Replace source files older than one week

+++ lib/php/FilesToJSON.php
@@ -123,7 +123,8 @@ class FilesToJSON
                 throw new \RuntimeException('Unknown type ' . $type);
         }
 
-        if (!file_exists($path)) {
+        $interval = strtotime('-1 week');
+        if (!file_exists($path) && filemtime($path) <= $interval) {
             file_put_contents(
                 $path,
                 file_get_contents($uri)
